TEST : loop server with TCP connect

+++ source/isocket/UDP_server.cpp
@@ -5,7 +5,7 @@
 // Login   <antoine.plaskowski@epitech.eu>
 // 
 // Started on  Fri Nov 20 07:53:57 2015 Antoine Plaskowski
-// Last update Fri Nov 20 15:37:45 2015 Antoine Plaskowski
+// Last update Mon Nov 23 09:54:36 2015 Anthony Bury
 //
 
 #include	<unistd.h>
@@ -90,16 +90,16 @@ uintmax_t	UDP_server::recvfrom(uint8_t &data, uintmax_t &size, std::string &clie
 	perror("inet_ntop()");
       else
 	client = ipv4;
-      return (ret);
+      return (static_cast<size_t>(ret));
     case AF_INET6:
       char      ipv6[INET6_ADDRSTRLEN];
       if (inet_ntop(AF_INET6, &sockaddr.ipv6.sin6_addr, ipv6, sizeof(ipv6)) == NULL)
 	perror("inet_ntop()");
       else
 	client = ipv6;
-      return (ret);
+      return (static_cast<size_t>(ret));
     }
-  return (ret);
+  return (static_cast<size_t>(ret));
 }
 
 uintmax_t	UDP_server::sendto(uint8_t const &data, uintmax_t &size, std::string const &client) const
